refactor : ContestTemplate 네이밍 통일
