+++ src/main/java/com/opus/opus/modules/contest/exception/ContestTemplateException.java
@@ -3,15 +3,15 @@ package com.opus.opus.modules.contest.exception;
 import com.opus.opus.global.base.BaseException;
 import com.opus.opus.global.base.BaseExceptionType;
 
-public class ContestTeamTemplateException extends BaseException {
-    private final ContestTeamTemplateExceptionType exceptionType;
+public class ContestTemplateException extends BaseException {
+    private final ContestTemplateExceptionType exceptionType;
 
-    public ContestTeamTemplateException(final ContestTeamTemplateExceptionType exceptionType) {
+    public ContestTemplateException(final ContestTemplateExceptionType exceptionType) {
         super(exceptionType.errorMessage());
         this.exceptionType = exceptionType;
     }
 
-    public ContestTeamTemplateException(final ContestTeamTemplateExceptionType exceptionType, final String message) {
+    public ContestTemplateException(final ContestTemplateExceptionType exceptionType, final String message) {
         super(message);
         this.exceptionType = exceptionType;
     }
@@ -21,4 +21,3 @@ public class ContestTeamTemplateException extends BaseException {
         return exceptionType;
     }
 }
-
